requirements.txt
Python==3.6.8
pip==18.1
Django==1.8.8
django-model-utils==3.1.1
pytz==2018.9
django-pagination==1.0.7
disqus==0.0.4


